# backend/swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/upload":{"post":{"operationId":"AppController_uploadFile","summary":"Upload an image","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}]}},"/files/{filepath}":{"get":{"operationId":"AppController_seeUploadedFile","summary":"Visualize uploaded file","parameters":[{"name":"filepath","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}]}},"/company-types":{"get":{"operationId":"AppController_getAllCompanyTypes","parameters":[],"responses":{"200":{"description":""}}}},"/article":{"get":{"operationId":"ArticleController_create","parameters":[],"responses":{"200":{"description":""}}}},"/mail":{"get":{"operationId":"MailController_create","parameters":[],"responses":{"200":{"description":""}},"tags":["mail"],"security":[{"bearer":[]}]}},"/user":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"responses":{"200":{"description":"users response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}},"201":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]}},"/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["user"],"security":[{"bearer":[]}]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"login response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"tags":["auth"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"responses":{"200":{"description":"login response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"tags":["auth"]}},"/auth/refresh-token":{"post":{"operationId":"AuthController_refreshToken","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"200":{"description":"login response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/request-reset-password-email":{"post":{"operationId":"AuthController_requestPasswordReset","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordReqDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/reset-password":{"post":{"operationId":"AuthController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/me":{"get":{"operationId":"AuthController_getAuthenticatedUser","parameters":[],"responses":{"200":{"description":"get authenticated user data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"tags":["auth"],"security":[{"bearer":[]}]}}},"info":{"title":"TUNLOG","description":"Tunlog apis","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ResponseDto":{"type":"object","properties":{"data":{"type":"object"}},"required":["data"]},"UserDispoDTO":{"type":"object","properties":{"goingTo":{"type":"string"},"startDay":{"type":"string"},"endDay":{"type":"string"},"startAt":{"type":"string"},"endAt":{"type":"string"},"comment":{"type":"string"}},"required":["goingTo","startDay","endDay","startAt","endAt","comment"]},"UserDTO":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"roleId":{"type":"number"},"companyName":{"type":"string"},"city":{"type":"string"},"country":{"type":"string"},"address":{"type":"string"},"websiteUrl":{"type":"string"},"commercialRegister":{"type":"string"},"patent":{"type":"string"},"companyTypeId":{"type":"number"},"userPackId":{"type":"number"},"carNumber":{"type":"string"},"carWidth":{"type":"number"},"carHeight":{"type":"number"},"carWeight":{"type":"number"},"disponibility":{"$ref":"#/components/schemas/UserDispoDTO"}},"required":["firstName","lastName","phone","email","password","carNumber","carWidth","carHeight","carWeight","disponibility"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"AuthResponseDto":{"type":"object","properties":{"accessToken":{"type":"object"},"refreshToken":{"type":"object"},"user":{"type":"object"}},"required":["accessToken","refreshToken","user"]},"RefreshTokenDto":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]},"ResetPasswordReqDto":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"ResetPasswordDto":{"type":"object","properties":{"newPassword":{"type":"string"},"resetPasswordToken":{"type":"string"}},"required":["newPassword","resetPasswordToken"]}}}}